November 2024 Enhancements (#58)
* November Updates

- Added expiring endpoint email notifications
- Added VLAN and dACL attributes
- Corrected bug when updating per user psk values

* New install detection

Detect new installation and provide setup guidance.

* DB Schema Installer Change

* Log Warning cleanup

* Modal show correction

Dependency messages not showing due to never correcting modal show logic during bootstrap major updates.

+++ adminportal/login.php
@@ -61,7 +61,10 @@
 
 	$samlSettings = $ipskISEDB->getGlobalClassSetting("saml-settings");
 	$samlLogin = (isset($samlSettings['enabled'])) ? $samlSettings['enabled'] : false;
-	$samlUsernameField = ($samlSettings['usernamefield'] != '') ? $samlSettings['usernamefield'] : 'REMOTE_USER';
+	
+	if ($samlLogin) {
+		$samlUsernameField = ($samlSettings['usernamefield'] != '') ? $samlSettings['usernamefield'] : 'REMOTE_USER';
+	}
 	
 	if ($samlLogin == true && $samlSettings['headers'] == true) {
 		$requestHeaders = getallheaders();
@@ -111,7 +114,7 @@
 								
 								$ipskISEDB->addUserCacheEntry($_SESSION['logonSID'],$_SESSION['userPrincipalName'],$_SESSION['sAMAccountName'],$_SESSION['logonDN'], $systemSID);
 								$adminSettings = $ipskISEDB->getGlobalClassSetting("admin-portal");
-								if($adminSettings['log-purge-interval'] != '') {
+								if(isset($adminSettings['log-purge-interval']) && $adminSettings['log-purge-interval'] != '') {
 									$ipskISEDB->purgeLogs($adminSettings['log-purge-interval']);
 								}
 								header("Location: adminportal.php");
